🚚 api: [CERTIF] move constants to certification/shared context
