+++ api/src/certification/shared/domain/constants/sessions-errors.js
